Add a unit test that proves out date range optimization.
Jeff noted that the FHIR REST layer in the engine was not using date
ranges to filter at the FHIR server and instead leaving that to be done
inside the engine. I spent some time debugging and found that the date
range logic only works when EnableDateRangeOptimization is enabled in
the CqlTranslator and then only when the CQL uses a well-known date path
in the retrieve operation. I created a unit test that proves it out.
Currently, though, the EnableDateRangeOptimization check is not
recommended because it can cause incorrect behavior when the date path
doesn't match a valid path.