Updates to SSL instructions
During QA verification it was discovered that the format of the keystore
generated by openjdk is pkcs12 by default (vs. JKS). Renamed the
trustStore.jks file to reflect the actual contents and updated the
instructions to include storetype both on truststore generation and
during JVM invocation so that it is crystal clear.