Code review response, refactoring library loading, callback enhancements
* There were a couple points in PR #9 that were addressed including
removal of extra semicolon, removal of dead file that should not have
been added, and an incorrectly resolved import for Base64 decoder.
* Renamed CqlTranslationWrapper to CqlTranslationProvider.
* Added some features to the evaluation callback around events for
entering and exiting a context. I think this will be helpful when we get
around to improving the result formatter in the main method.
* Added a ClasspathLibrarySourceProvider to be used in the unit tests
instead of force loading the libraries via the addLibrary method.
* Made a bunch of test changes to support the refactored library
management code. The changes exposed an issue with how we are loading
libraries. The test resources that were created do not follow the
prescribed name-version filename pattern, so we had to do some hacks to
make the tests work.