WFFHCOHORT-265 - Added/verified UI zip file support
The UI ZIP files are multi-leveled and contain extra things that are not
FHIR resources. I added some filtering capability to the existing ZIP
support in the CohortCLI in order to deal with that structure and to
generally give the user finer grained control of what gets considered.
In the MeasureCLI, there was no prior support so that took a bit of
overhauling. I updated the "-m" flag to the CLI to support a directory,
a zip, or the prior path of direct REST calls to a FHIR server.
Directory wasn't technically part of this ticket, but it was trivial to
add after the infrastructure was in place, so I added it. I also added
bonus support for FHIR Bundle resources in either zip or directory
loading mechanisms in case we decide we want to compile the resources
instead of keeping them as separate files. It would be nice to go back
and overhaul the CohortCLI to support the same interpretation of the -m
flag, but that seems overly ambitious for this ticket and generally
unnecessary. This ticket also raised the idea of what we would do for
filesystem or ZIP support for the data and terminology needs of the
client. It is a little silly to only have partial support. I will log
another ticket for that work at some later point.

As a note, the FHIR resources were modified in anticipation of upcoming
changes to the UI export that are not yet complete. This work could
change a little after that is finalized and delivered.