WFFHCOHORT-235 add support for executing measure by canonical URL
This support was largely already in place just using the measure ID
field in the JSON input or the -r flag on the command line and providing
the canonicalURL instead of the id. Our handling of version was
inconsistent across the different loading mechanisms, though, so I
overhauled that to leverage the infrastructure that Matt already put in
place for lookups by identifier. I also propagated the pattern to the
RestFhirLibraryResourceProvider for consistency and to reduce
duplication of concerns. A lot of the REST implementation overlaps with
the filesystem/zip implementation of resource loading that was added in
WFFHCOHORT-265, but I am keeping them separate for now because I believe
the caching concerns are somewhat different based on where the data is
coming from.