Overhaul importer to support updated export artifact.
This is a major change to how the importer code works. Because the new
UI bundles are now using reliable references between artifacts, the
importer is able to follow a much simpler and more atomic path towards
getting the resources loaded in the server. All resources are added to a
bundle and then processed as a single transaction.

While I was working on this I had a long conversation with the IBM FHIR
team on how to approach concurrent updates where two measure bundles
might be trying to create the same Library resource. The only reliable
way to deal with this in IBM FHIR is to provide a resource ID. The UI
artifacts want to use the name + version as the logical key and do not
provide a resource ID. To guarantee only a single resource is ever
created in FHIR, a resource ID generation scheme was created. The
default scheme uses SHA-256 hashing and Base64 encoding as the basis for
the ID generation with some permutations applied to match the
constraints specified by FHIR for Resource.id. The resulting IDs are not
absolutely 100% guaranteed to be unique, but practically speaking they
will be unique.

The HAPI FHIR client will throw an exception if a transaction bundle
fails to complete. Rather than crater the entire process, those errors
are now captured and the result code of the JVM execution is the number
of errors that occurred during processing. This will allow a calling
script to easily check the result code and take action as needed.

Because much of the work is now done in bundles, the importer will write
the request and response to an output folder so that users can examine
the results as needed.