Address "Resource for null/XXX" ugliness in MeasureCLI output.
When running the MeasureCLI using either the ZIP or Filesystem resource
providers, the output of the MeasureCLI included the message "Resource
for null/XXX" where XXX is the measure ID. This was related to the fact
that the UI resources don't include an ID value and instead we create
one. We weren't setting the resource type in the ID, so it came out as
null. I updated the provider code to include the resource type and added
checks to the test cases to make sure no nulls appear in the output.

During the builds, I noticed a few javadoc warnings that I also
addressed in CanonicalHelper and MeasureResolutionProvider.