Address double-hit to the FHIR server for library resolution.
Kwas noticed that after my recent PR merge that we were starting to get
double-hits to the FHIR server during the Library load step of measure
evaluation. The issue has to do with a change in my PR to proactively
load all of the FHIR Library resources in the dependency graph prior to
trying to do the measure evaluation. Previously, only the top level was
loaded and then lazy loading was used to bring in the rest of the
Library resources using the CqlEngine LibraryLoader interface.

If the measure references are by URL, there is a cache miss when the
CqlEngine LibraryLoader tries to resolve by name and version. I already
had some logic in place to keep the three independent lookup caches in
sync when the entry point was the ID, but I didn't have it in place for
the by URL entry point. I updated the code so that all three caches are
kept in sync regardless of what entry point is used.