Fix inexact name matching in RestFhirLibraryResolutionProvider
The RestFhirLibraryResolutionProvider was using matches() in name-based
queries and that is problematic because that translates to a ``name=``
URL parameter which does a "contains" check in the FHIR server. I
updated the resolver to use matchesExactly (aka name:exact=) in order to
address issues where test resources have similar names.