WFFHCOHORT-178 Initial implementation of the evaluation API.
* Implementation assumes a multipart form request that consists of a
JSON structure that describes the evaluation to perform and a ZIP
artifact that contains the FHIR resources for the needed Measure and
Library artifacts. Patient data is assumed to be in an actual
FHIR server and the connection to that FHIR server is described in the
metadata part of the request. This supports testing where the Measure
and Library resources need to be iterated without modifying the
published versions on the FHIR server.
* Deployment is assumed to be Liberty (IBM or Open) and thus the Liberty
APIs for multipart requests are used.
* Requests are for a single patient and single measure at a time. All of
the previous work that was done to support a batch-style interface was
removed in order to have a clean API that only describes the implemented
functionality. The batch interface may come back in a future release.
* A lot of classes were added to support passing parameter overrides in
the metadata part of the request. These classes will probably migrate to
the cohort-engine project, but we need to resolve the function here
against what is in the CLI project first.