Move the Parameter objects from api.model to cohort_engine
After I added the Parameter objects to api.model to handle input from
the REST layer in String form, there was a duplication of concern with
the handling in the CLI project. As a solution, I moved the parameter
objects into the cohort-engine project and then made code updates so
that the APIs and CLIs would use the new Parameter objects. This had a
fairly wide ranging impact.

While I was making this change, I noticed that the API for
MeasureEvaluation did not support evaluation by business identifier and
version like other APIs. In order to support this type of lookup, I
switched the model object over to use the MeasureContext object that was
already in place for CLI evaluation.

Tim mentioned that we might want to dark launch the evaluation API so
that we wouldn't accidentally use it in non-dev environments, so I added
the necessary boilerplate around the REST API to support feature
enable/disable functionality via the dark launch framework.