Further refactoring so that CQLEngineWrapper takes Parameter objects in
When I did the update to support Parameter objects, I carried it through
as far as CohortCLI, but left the CqlEngineWrapper alone. This update
continues with Parameter objects all the way into the CqlEngineWrapper.

I also moved the new Jackson helpers to a new package in the cohort-util
project because I can see them being useful elsewhere.