Leveraging Caffeine's use of Lightbend config
This comes with a move towards letting the JCache provider decide how they are to be configured instead of having each application require the same parameters to configure the cache each time.
This also simplifies the CLI parameter footprint of both `MeasureCLI` and `CohortEngineFlinkDriver`.

I was forced to add a merging step to the uberjar process to retain the contents of Caffeine's `reference.conf` file.