Add unit tests for the unsupported ValueSet features
Jai noticed the UnsupportedOperationException wasn't being generated in
the filtered retrieve case (e.g. [Condition:MyValueSet]). I added these
unit tests to verify the behavior. It turns out that there is additional
work needed in the OSS project in order to enable these properties to
even be visible to the terminology provider during ValueSet expansion
during the filtered retrieve. That is work that will require changing
internal classes and an interface definition, so it isn't easily patched
on our side. I documented the issue in the OSS project
(https://github.com/DBCG/cql_engine/issues/461#issuecomment-816924813)
and we can pursue it further with them. I left the tests in place but
marked them ignore for now.