Fix resolveMeasureByIdentifier issue.
When I brought over the resolveMeasureByIdentifer method into the
ResourceResolutionProvider, I broke it. I added some unit tests to
verify behavior and then rewrote the method to work correctly.

I noticed while I was doing this that the implementations of the
ResourceResolutionProvider (directory and zip) were in different
packages. Directory was in the cohort.engine package and the base class
and zip implementation were in the (cohort.engine.measure) package. This
was not intentional. I moved the directory implementation to the same
package as the other two.