Add unit test for "interval starts interval"
It was noted that the CQL Author's Guide lists "interval begins
interval" as a valid operator, but it wasn't functioning in the engine.
After investigation, it was determined to be a documentation problem.
The correct operation is "interval starts interval". This unit test
validates that operator exists and is functional.