Patch the DataProvider to handle :in queries properly.
There is an option on the TerminologyAwareDataProvider that allows it to
build FHIR search queries that use the token :in modifier to do
terminology expansion on the FHIR server during search. Use of this
feature has the potential to speed up performance by reducing round
trips to the FHIR server to get value set expansions and reducing round
trips to the search API for value sets with a large number of codes.

Under testing it was found that the OSS implementation that builds the
REST calls is broken. This appears to be an issue in the HAPI client
library. Prior to the use of the HAPI client to build queries, there was
a functional implementation in SearchParameterMap. That implementation
still exists and I added a code path that will fall back to it when any
modifier is found in the search parameters.

Since I was already patching the DataProvider, I added support for
configuring the pageSize requested from the FHIR server during search
requests. This will allow the user to override the very small default
values found in IBM and HAPI servers. Note that this is now configurable
on the server side in IBM FHIR 4.7.0. I set a default page size of 1000
because that is the maximum allowed in IBM FHIR.

I had to touch the CohortEngineRestHandler to add support for the new
configuration parameters, so I patched the documentation while I was in
there to provide a complete example of how to invoke the evaluation API.