remove uneeded file
Per Andrew G: we have a check in the CT pipeline that will look if you
have and COS buckets in your account and then check if they are
encrypted.  If any of thosebuckets are for POC work, or other reasons
that they should not be checked, that file is used.  So yes, that file
is NOT relevant to your team and can be removed.