Enhance the logging around library loading and translation
While doing some debugging with QA on an issue related to CQL
translation, we found that the detail provided to clients in our CQL
Translation errors was inadequate to isolate the problem. The changes in
this PR help narrow in on the problem. Specifically, it adds to
translation errors the library name, library version, and start and end
positions of the expression that caused the failure. Logging was also
added for translation warnings and library loading.