API updates for CqlEvaluator and ELM deserialization
While working with the health patterns team, I found that the
CqlEvaluator interface was inconsistent with regard to the APIs exposed
for CqlEvaluationResult and CqlExpressionResult callbacks. I added a few
signatures so that we have consistency.

While I was working, I noticed that the split of the
CqlTranslationProvider into its own project moved some code into the
translator project that did not technically belong to the translator.
The translator produces ELM in an execution agnostic ELM format, the CQL
Engine consumes that ELM and maps it into an execution typesystem. The
custom object factory code is only needed on the execution side and not
the translation side. I moved the OptimizedCqlLibraryReader and its
related classes back into the cohort-engine project where they started.
This means that the translation component no longer has the ability to
render the ELM XML in the execution typesystem. I had to change the
CqlTranslationProvider interface to produce strings instead of inflated
Library objects to make this work. There aren't any preexisting
consumers consuming translation directly, so this should be fine. There
is probably more room for improvement here (perhaps the optimized
Library reader code could be its own module), but it is better now than
it was when forcing the translator dependency on the runtime just to
deserialize the ELM.