Switch DataRowRetrieveProvider to use Code equivalence checking (#173)
* Update RetrieveProvider to use equivalence-based code checks.

For retrieves that are filtered by codes, the DataRowRetrieveProvider
indexes the codes in the _codePath_ column on first access. The CodeKey
object is used as the index key. CodeKey extends the CQL Code object and
implements hashCode and equals so it works as a Map key. It *was* using
the the CQL Engine's Code.equal method for equality checks. Code.equal
takes into account all fields in the Code including version and display.
Per the CQL spec, the retrieve filtering should use equivalence checking
instead. I updated CodeKey.equals to use Code.equivalent and fixed the
hashCode method so that it only considers the system and code fields.

I made a small change to the DataRowRetrieveProvider to handle null
values and to allow users to specify any type of column in codePath. If
the column is non-null and not already a Code, a Code will be created
and the value will be set to the toString of the codePath object.

The DataRowRetrieveProviderTest was refactored to avoid a bunch of
unnecessary duplication of test data initialization. I then fleshed out
the tests so that we have complete code coverage of all the code paths.

* Add a unit test for SparkDataRow when metadata has default system.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>
Signed-off-by: TimBishop <tabishop@us.ibm.com>