Fix expression language classpath issue in Spark App
Spark includes hadoop-common which includes the JSP 2.1 API JAR. The JSP
2.1 API JAR has some javax.el API classes in it that are not compatible
with modern implementations of the EL classes. I had a choice to either
backlevel our javax.validation dependency to 4.2.1.Final and javax.el to
1.0.0.GA or increment our JSP version to 2.2. I went with the later.
This required an update to the Dockerfile to remove the JSP API version
included with the Spark distribution and some dependency management to
make sure we had the necessary JARs in the Spark application classpath.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>