WFFHCOHORT-627 Initial implementation of column filtering
* Implemented a dataset filtering mechanism that uses information
collected using the DBCG ElmREquirementsVisitor to determine which
columns are use in the libraries under evaluation are needed and then
selects only those columns during the read dataset phase of the program.
I couldn't make the expression-level filtering of dataset requirements
function correctly using this code, so requirements filtering is done at
the full library level.
* In order to use the DBCG code, it was necessary to update our CQL
translator dependency to version 1.5.4. This caused some ripple down
effects into the cql-translation and cql-engine projects, as well as,
the SparkSchemaCreator.
* Added missing copyright headers to a bunch of files.
* Removed the unused and poorly implemented externalId field from the
CqlLibraryDescriptor object.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>