Fix failure caused by FHIRHelpers with missing EnableResultTypes option.
The ELM visitor logic that collects the paths that are utilized in the
logic for each datatype relies on the EnableResultTypes option being
used using CQL translation. The FHIRHelpers objects are available in the
classpath in pre-compiled form that did not use that option, so they
don't work for the path capture operation. I added logic to validate
that the EnableResultTypes option was used during translation and
updated our use of classpath resources to only accept CQL resources vs.
either CQL or ELM. This forces use to compile FHIRHelpers for our use.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>