WFFHCOHORT-627: Add intelligent column filtering to spark reads (#188)
* WFFHCOHORT-627 Initial implementation of column filtering

* Implemented a dataset filtering mechanism that uses information
collected using the DBCG ElmREquirementsVisitor to determine which
columns are use in the libraries under evaluation are needed and then
selects only those columns during the read dataset phase of the program.
I couldn't make the expression-level filtering of dataset requirements
function correctly using this code, so requirements filtering is done at
the full library level.
* In order to use the DBCG code, it was necessary to update our CQL
translator dependency to version 1.5.4. This caused some ripple down
effects into the cql-translation and cql-engine projects, as well as,
the SparkSchemaCreator.
* Added missing copyright headers to a bunch of files.
* Removed the unused and poorly implemented externalId field from the
CqlLibraryDescriptor object.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Implement fail-safe control for data filtering

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Additional unit tests for data filter functionality

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Add short-circuit logic to remove dataset reads for unused data.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Tweaking the unit test to verify indirect data reqs in grandchild

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Merge in Tim's changes and fix some whitespace issues

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Add support for AnyColumn filtering

* Added a new ELM visitor that looks for AnyColumn function usage and
reports the matching logic used by those functions for use in downstream
column filtering.
* Updated the DataTypeRequirements to gather the AnyColumn requirements
as a list of StringMatcher functions that can be applied in the
FilteredDatasetRetriever.
* Updated the SparkCqlEvaluator and FilteredDatasetRetriever classes to
use StringMatcher logic.
* In an effort to avoid warnings on the AnyColumn method, I renamed the
containing class to AnyColumnFunctions so AnyColumn doesn't look like a
constructor. I also added missing copyrights and updated the methods to
use the new StringMatcher utility classes that I created.
* Added other missing copyrights and SuppressWarnings messages to try to
clean up my problems view in Eclipse

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Respond to simplest code review items

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Switch to custom ELM visitor for capturing data requirements

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Fix failure caused by FHIRHelpers with missing EnableResultTypes option.

The ELM visitor logic that collects the paths that are utilized in the
logic for each datatype relies on the EnableResultTypes option being
used using CQL translation. The FHIRHelpers objects are available in the
classpath in pre-compiled form that did not use that option, so they
don't work for the path capture operation. I added logic to validate
that the EnableResultTypes option was used during translation and
updated our use of classpath resources to only accept CQL resources vs.
either CQL or ELM. This forces use to compile FHIRHelpers for our use.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Add propagation of base-class properties to child classes

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Add some unit tests that exercise the column filtering program arg

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Fix test method names to reflect their actual behavior

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Fix errors reported by newly added PMD rule

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Respond to code review feedback

* Switch from Stack to Deque
* Add more trace logging for consistency across ELM nodes
* Clean up and guard some existing trace logging for best runtime
performance

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>

* Add test case for requesting a column not in the dataset

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>
Signed-off-by: TimBishop <tabishop@us.ibm.com>
Signed-off-by: David Kwasny <dkwasny@us.ibm.com>