Enable comma-separate value parsing of list args.
Code review noted that one of the new parameters added for output result
grouping was not plural. It was intentionally defined in singular form
because it was assumed that each instance of the parameter was a
singular value. Doing some more investigation into the JCommander
documentation, I found that it is possible out-of-box to get list-value
parsing of comma-separated values, but it requires that the parameter
type be a List. Set parameters will not receive the special argument
parsing.

I switched all the Set parameters to List parameters and updated some of
the long names of the parameters to plural form to reflect that they can
be multi-valued (e.g. expressions, aggregation-contexts,
key-parameters). I added some unit tests to confirm the argument parsing
is working as expected.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>