WFFHCOHORT-689 - Split output into multiple rows based on CQL parameters (#208)
* WFFHCOHORT-689 - Split output into multiple rows based on CQL parameters

Instead of every evaluation result per context being joined into one
large row, consumers have asked that we split related evaluation results
into separate rows. For example, if the same logic is evaluated for
multiple time ranges, the output will contain a column for each piece of
logic and a row for each time range.

Rows are split based on a grouping key. The grouping key consists of one
or more CQL parameter values. By default all parameter values are used.
In the case where there are more CQL parameter values than are needed
for the grouping key, a new program argument `--key-parameters` is
provided. This parameter can be repeated as many times as needed to
define the exact set of parameters that define the grouping key. The
behavior can be disabled by using the `--disable-result-grouping`
argument.

Signed-off-by: Corey Sanders <corey.thecolonel@gmail.com>
Signed-off-by: TimBishop <tabishop@us.ibm.com>
Signed-off-by: David Kwasny <dkwasny@us.ibm.com>